Dependency errors are so fun!
Interestingly loading anything that has org-mode as a dependency
before loading org itself, will cause it to try to use the built-in
org package from the emacs installation, even if there is a new one
installed, this of course makes the new one also not work when it
tries to load, since the old one was already loaded, so fun!

This seems to be an inherent problem with straight.el because it
doesn't really load any packages until you specify them in your init
file.